pyproject.toml
[tool.poetry]
name = "swapi-dataplatform"
version = "0.1.0"
description = ""
authors = ["robgriffin247 <robgriffinwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python =  ">=3.10,<3.11"
requests = "^2.31.0"
snowflake-snowpark-python = {extras = ["pandas"], version = "^1.10.0"}
dbt-snowflake = "^1.7.0"
python-dotenv = "^1.0.0"
path = "^16.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
